refactor: isolate deepseek streaming aggregator
